+++ src/module/effects/effect-models.ts
@@ -1,7 +1,0 @@
-export enum EffectActions {
-  create = 'create',
-  edit = 'edit',
-  delete = 'delete',
-  toogle = 'toggle',
-  update = 'update',
-}